necktie/requirements.txt
django==3.2.15
django-filter==21.1
drf-yasg==1.20.0
djangorestframework==3.12.4
# Postgres
psycopg2==2.9.2
parse==1.19.0
